fix: decode no longer panics on u16-overflowing span length (#33)
decode computed a span's run length as (right - left + 1) in u16. A
header with left=0, right=65535 overflows that u16 (65535 + 1) and
panics under overflow checks, before the widen to usize — breaking the
documented no-panic contract on untrusted wire input. Widen the
operands first; the right >= left invariant just above keeps the usize
subtraction from underflowing.

Found by the cargo fuzz serialize lane (~0.2s); the robustness proptest
did not reach this input. Adds a regression test decoding a crafted
left=0, right=65535 buffer and asserting a typed Err, not a panic.

+++ src/serialize.rs
@@ -164,7 +164,11 @@ pub fn decode(bytes: &[u8]) -> Result<Frame, DecodeError> {
         if right < left {
             return Err(DecodeError::BadSpan);
         }
-        let n = (right - left + 1) as usize;
+        // Widen before the arithmetic: `right - left + 1` in `u16` overflows
+        // when `right == u16::MAX` (e.g. left=0, right=65535), panicking under
+        // overflow checks. `right >= left` is enforced just above, so the
+        // subtraction in `usize` cannot underflow.
+        let n = right as usize - left as usize + 1;
         let mut cells = Vec::with_capacity(n);
         for _ in 0..n {
             cells.push(decode_cell(&mut r)?);
